Range of g-weights

Computes the range of the ratios between calibrated weights and direct weights (g-weights) for the original sample and all its replicates.

g.range(cal.deskott)

Arguments

cal.deskott
Object of class kott.cal.design.

Details

This function computes, for the original sample and all its replicates, the smallest interval which contains the ratios between calibrated weights and direct weights.

Value

A data.frame object.

See also

Examples

data(data.examples)

# Creation of a kott.design object:
kdes<-kottdesign(data=example,ids=~towcod+famcod,strata=~SUPERSTRATUM,
      weights=~weight,nrg=15)

# Calibration:
kdescal04p<-kottcalibrate(deskott=kdes,df.population=pop04p,
            calmodel=~x1+x2+x3-1,partition=~regcod,calfun="logit",
            bounds=bounds,aggregate.stage=2)

# Computation of g-weights range:
g.range(kdescal04p)

#>          sample     g.min    g.max
#> 1      original 0.7074003 1.700717
#> 2   replicate.1 0.6938845 1.608282
#> 3   replicate.2 0.6386560 1.695114
#> 4   replicate.3 0.6405845 1.650475
#> 5   replicate.4 0.6835884 1.953700
#> 6   replicate.5 0.7009600 1.633713
#> 7   replicate.6 0.6530224 1.769620
#> 8   replicate.7 0.7513285 1.428188
#> 9   replicate.8 0.6714811 2.084896
#> 10  replicate.9 0.6968937 1.484918
#> 11 replicate.10 0.7252601 2.159878
#> 12 replicate.11 0.6199823 1.623145
#> 13 replicate.12 0.6654837 1.652526
#> 14 replicate.13 0.7071155 1.647926
#> 15 replicate.14 0.7018868 1.652705
#> 16 replicate.15 0.6709386 1.731815